% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mindware_sample_study_1}
\alias{mindware_sample_study_1}
\title{Mindware Sample Study}
\usage{
mindware_sample_study_1()
}
\value{
path to sample MindWare study; character string
}
\description{
Mindware Sample Study
}
