% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook.R
\name{MW}
\alias{MW}
\title{Read a MindWare Workbook}
\usage{
MW(path)
}
\arguments{
\item{path}{a character string; path to a workbook}
}
\value{
a list of data frames; psyphr workbook S3 object
}
\description{
Read a MindWare Workbook
}
