% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_study_mindware}
\alias{sample_study_mindware}
\title{Mindware Sample Study}
\usage{
sample_study_mindware()
}
\value{
path to sample MindWare study; character string
}
\description{
Mindware Sample Study
}
