% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\title{Read a study}
\usage{
study(path, ...)
}
\arguments{
\item{path}{path to a study directory; a character string}

\item{...}{dot-dot-dot}
}
\value{
a data frame; psyphr study S3 object
}
\description{
Read a study
}
