% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{unnest_data}
\alias{unnest_data}
\title{Unnest Data in a Study}
\usage{
unnest_data(study)
}
\arguments{
\item{study}{a psyphr study}
}
\value{
a data frame; psyphr study S3 object
}
\description{
Unnest Data in a Study
}
